+++ e2e/packages/sync-test/data/expectClientData.ts
@@ -1,0 +1,13 @@
+import { Page, expect } from "@playwright/test";
+import { Data } from "./testData";
+import { readClientStore } from "./readClientStore";
+import config from "../../contracts/mud.config";
+
+export async function expectClientData(page: Page, data: Data) {
+  for (const [table, records] of Object.entries(data)) {
+    for (const record of records) {
+      const value = await readClientStore(page, [config.namespace, table, record.key]);
+      expect(value).toEqual(record.value);
+    }
+  }
+}
