+++ e2e/packages/sync-test/data/index.ts
@@ -1,0 +1,4 @@
+export * from "./expectClientData";
+export * from "./readClientStore";
+export * from "./setContractData";
+export * from "./testData";
