+++ e2e/packages/sync-test/setup/startViteServer.ts
@@ -1,0 +1,13 @@
+import { ViteDevServer, createServer } from "vite";
+
+export async function startViteServer(): Promise<ViteDevServer> {
+  // TODO this should probably be preview instead of dev server
+  const mode = "development";
+  const server = await createServer({
+    mode,
+    server: { port: 3000 },
+    root: "../client-vanilla",
+  });
+  await server.listen();
+  return server;
+}
